+++ example/lib/homeview.dart
@@ -7,37 +7,49 @@ class HomeView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    const duration = Duration(seconds: 120);
+    const duration = Duration(seconds: 5);
+    final cache = CacheData<List<Data>>();
+
     return Scaffold(
-      body: FutureBuilder<List<Data>?>(
-        future: CacheData<List<Data>>().fetchData(
-            url: 'https://jsonplaceholder.typicode.com/albums',
-            duration: duration),
-        builder: (context, snapshot) {
-          if (snapshot.hasError) {
-            // throw snapshot.error!;
-            print(snapshot.error!);
-            return const Center(
-              child: Text('An error has occurred!'),
-            );
-          } else if (snapshot.hasData) {
-            return ListView.builder(
-              itemCount: snapshot.data!.length,
-              itemBuilder: (context, index) {
-                return ListTile(
-                  leading: CircleAvatar(
-                    child: Text(snapshot.data![index].userId.toString()),
-                  ),
-                  title: Text(snapshot.data![index].title),
-                );
-              },
-            );
-          } else {
-            return const Center(
-              child: CircularProgressIndicator(),
-            );
-          }
+      body: RefreshIndicator(
+        onRefresh: () async {
+          cache.fetchData(
+              'https://jsonplaceholder.typicode.com/albums',
+              duration: duration);
+          debugPrint("isExpired ${cache.isExpired(
+            'https://jsonplaceholder.typicode.com/albums',
+          )} now ${DateTime.now()}");
         },
+        child: FutureBuilder<List<Data>?>(
+          future: cache.fetchData(
+              'https://jsonplaceholder.typicode.com/albums',
+              duration: duration),
+          builder: (context, snapshot) {
+            if (snapshot.hasError) {
+              // throw snapshot.error!;
+              debugPrint(snapshot.error!.toString());
+              return const Center(
+                child: Text('An error has occurred!'),
+              );
+            } else if (snapshot.hasData) {
+              return ListView.builder(
+                itemCount: snapshot.data!.length,
+                itemBuilder: (context, index) {
+                  return ListTile(
+                    leading: CircleAvatar(
+                      child: Text(snapshot.data![index].userId.toString()),
+                    ),
+                    title: Text(snapshot.data![index].title),
+                  );
+                },
+              );
+            } else {
+              return const Center(
+                child: CircularProgressIndicator(),
+              );
+            }
+          },
+        ),
       ),
     );
   }