+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## 0.0.2
-- Update README.md
 ## 0.0.1
 - First release
+
+## 0.0.2
+- Update README.md
+
+## 1.0.0
+- refactor: remove re-instantiate http client from every method
+- refactor: remove re-instantiate [CacheRepository] class from every method
+- feat: introduce delete item from cache
+- feat: introduce delete all items from cache
+- feat: introduce expired status of cache